moving assertion files into folder Assertion
